+++ shiny/experimental/ui/_htmldeps.py
@@ -23,9 +23,9 @@ def card_full_screen_dep() -> HTMLDependency:
 
 def fill_dependencies() -> HTMLDependency:
     return HTMLDependency(
-        "htmltools-fill-x",
+        "htmltools-fill",
         "0.0.0.0",
-        source={"package": "shiny", "subdir": str(ex_www_path)},
+        source={"package": "shiny", "subdir": str(ex_www_path),},
         stylesheet={"href": "fill.css"},
     )